cositas css en nav bar y search
