+++ henryproject/Client/src/Components/Games/Games.css
@@ -31,6 +31,7 @@
 
 .Search-Filters{
     margin-top: 10%;
+    min-height: 100%;
 }
