+++ henryproject/Client/src/Components/Nav_bar/Nav_bar.css
@@ -56,7 +56,7 @@
 .Nav_bar {
     display: flex;
     background-color: #3131317F ;
-    position: absolute;
+    position: fixed;
     z-index: 10;
     top: 2%;
     right: 2%;
