+++ henryproject/Client/src/Components/Nav_bar/Nav_bar.jsx
@@ -1,24 +1,28 @@
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import React from "react";
 import Icon from '../../Style/Imagenes/Icon.PNG'
 import Cart from '../../Style/Imagenes/cart.png'
 import Cora from '../../Style/Imagenes/Corazon.png'
 import { useSelector } from 'react-redux'
 import SignUserModal from './SignUserModal.jsx'
+import SearchBar from '../SearchBar/SearchBar'
 
-import UserPop from './UserPop.jsx'
+//import UserPop from './UserPop.jsx'
 
 import './Nav_bar.css'
 import { useState } from "react";
 import ProfileNav from "../ProfileNav/ProfileNav";
-const axios = require('axios')
+//const axios = require('axios')
+import { getAllGames } from '../../redux/Actions/Index';
+import { useDispatch } from 'react-redux';
 
 
 export default function Nav_bar({ userLogged, setUserLogged }) {
 
     const itemsCart = useSelector(state => state.cart)
     const itemsFavorites = useSelector(state => state.favorites)
-
+    const dispatch = useDispatch()
+    const navigate = useNavigate()
 
 
     const [isOpen, setIsOpen] = useState(false);
@@ -27,6 +31,12 @@ export default function Nav_bar({ userLogged, setUserLogged }) {
         setIsOpen(!isOpen);
     }
 
+    const onSearch = (name) => {
+        navigate("../home/games", { replace: true });
+        dispatch(getAllGames(name))
+    }
+
+
     return (
         <nav className="Nav_bar">
 
@@ -34,9 +44,18 @@ export default function Nav_bar({ userLogged, setUserLogged }) {
                 <img src={Icon} alt="Icon" />
             </Link>
 
+            <div>
+                <SearchBar
+                    onSearch={onSearch}
+                ></SearchBar>
+            </div>
 
             <div>
                 <Link to='/home'><button>Home</button></Link>
+            </div>
+
+            <div>
+                <Link to='/home/games'> <button> Search in our game list </button> </Link>
             </div>
             {/* <Link to='/home/create'><button>Create Videogame</button></Link>
